Assistant checkpoint: Add PWA meta tags and fix manifest start URL
Assistant generated file changes:
- templates/index.html: Add manifest link and mobile viewport tags
- manifest.json: Update manifest start URL

---

User prompt:

I would like my app to not behave like a website when used on mobile, but I would like to be able to save it as a bookmark and then it behaves like an app when opening PWA. I added a file called manifest.json to the home directory and I added a banner_app.png picture to the static folder. In another project you told me these are the prerequisites. is this correct? anything missing? I do not want to add offline functionality. please check, but do not impacet the general functionality of the app with any changes