Assistant checkpoint: Update mobile layout for better usability
Assistant generated file changes:
- static/style.css: Update mobile input area layout
- templates/index.html: Update input area structure

---

User prompt:

i deployed the app. on the iphone there is now some blank space between the text input window at the bottom and the text window where the messages appear. we could use this space in the following way:
Put the input field into a separate line as at the moment it is quite narrow - and below have the buttons for file upload, camera, send (file upload and camera buttons should be ond the left, Send button should be on the right. Then we can also extend the output window to make it longer. is it possible to make it flexible so it adjust to any screen size? I am only referring to the mobile version now. please make a suggestion.