Assistant checkpoint: Add prompt composer functionality
Assistant generated file changes:
- templates/index.html: Add prompt composer HTML
- static/style.css: Add prompt composer styles
- static/script.js: Add prompt composer functions

---

User prompt:

At the moment the input field for the user messages is a simple field. I want to give the user the opportunity to submit messages in a more structured way. the idea is that there will be a button at next to the send button which will open a "prompt composer" window. it should slide in from the right. This window should have five text entry fields: Instructions, Return Format, Context, Examples, Warnings. The user can enter text into each of the fields. At the bottom should be a "Transfer" button which then transfers a concatenation of the content of all the fields into the regular text entry field. In the concatenated string the content of each of the text fields should be wrapped in tags for easy identification by the LLM. If a field is empty, no tags should be included in the string. Below are the tags of the different components. After clicking the transfer button, the composer window should close. Additionally there should be a clear button to clear all the fields.

<instructions>
</instructions>

<format>
</format>

<context>
</context>

<examples>
</examples>

<warnings>
</warnings>