Assistant checkpoint: Update web search toggle to match button style
Assistant generated file changes:
- templates/index.html: Replace search checkbox with styled button
- static/style.css: Add styles for search toggle button
- static/script.js: Add web search toggle function, Update search enabled check in fetchChatResponse

---

User prompt:

can we change the design of the "Enable Web-Search" toggle? We should use a  button with a symbol from the same family as the buttons that we are already using for camera, upload, etc. Maybe a "globe" symbol would fit very well. If it is de-activated it should have a grey background, if it is activated it should have the usual blue background. default should be de-activated (grey)