Assistant checkpoint: Update citation display and styling
Assistant generated file changes:
- static/script.js: Update citation formatting and display
- static/style.css: Update citation styles

---

User prompt:

is this how the response is supposed to look? I do not see any differently formated links at the bottom of the message:

Based on the provided search results, I'll use the most timely and relevant ones to answer the user's question "test".

The top results suggest that the user might be looking for information on taking a test, testing speed, or the definition of the word "test".

Here are a few possible answers:

1. If you're looking to take a test, you can visit [https://implicit.harvard.edu/implicit/takeatest.html](https://implicit.harvard.edu/implicit/takeatest.html) to take an Implicit Association Test (IAT) on various topics.
2. To test your internet speed, you can use [https://www.speedtest.net/](https://www.speedtest.net/) or [https://fast.com/](https://fast.com/).
3. If you're looking for the definition of the word "test", you can check [https://www.merriam-webster.com/dictionary/test](https://www.merriam-webster.com/dictionary/test) or [https://www.dictionary.com/browse/test](https://www.dictionary.com/browse/test).

Please let me know if you need more specific information or clarification on any of these options!