Assistant checkpoint: Added Excel file extensions to file input accept attribute
Assistant generated file changes:
- templates/index.html: Add Excel file types to the file input accept attribute

---

User prompt:

this seemed to have worked - just a little thing:
When clicking the file upload button and selecting a directory, I am not seeing the excel files for selection. I manually need to switch the file type to "all files" to see it. Then i can upload it and it works. Did we forget something here?